+++ tycho-artifactcomparator/src/main/java/org/eclipse/tycho/zipcomparator/internal/ManifestComparator.java
@@ -14,6 +14,8 @@ package org.eclipse.tycho.zipcomparator.internal;
 
 import java.io.IOException;
 import java.util.LinkedHashSet;
+import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Set;
 import java.util.TreeMap;
 import java.util.jar.Attributes;
@@ -24,11 +26,16 @@ import org.codehaus.plexus.component.annotations.Component;
 import org.eclipse.tycho.artifactcomparator.ArtifactComparator.ComparisonData;
 import org.eclipse.tycho.artifactcomparator.ArtifactDelta;
 import org.eclipse.tycho.artifactcomparator.ComparatorInputStream;
+import org.osgi.framework.Constants;
 
 @Component(role = ContentsComparator.class, hint = ManifestComparator.TYPE)
 public class ManifestComparator implements ContentsComparator {
 
     public static final String TYPE = "manifest";
+
+    private static enum Change {
+        BASELINE_ONLY, PROJECT_ONLY, DIFFERENT;
+    };
 
     private static final Set<Name> IGNORED_KEYS = Set.of(
             // these keys are added by plexus archiver
@@ -48,41 +55,81 @@ public class ManifestComparator implements ContentsComparator {
     @Override
     public ArtifactDelta getDelta(ComparatorInputStream baseline, ComparatorInputStream reactor, ComparisonData data)
             throws IOException {
-        TreeMap<String, ArtifactDelta> result = new TreeMap<>();
+        TreeMap<String, ManifestDelta> result = new TreeMap<>();
 
-        Manifest manifest = new Manifest(baseline);
-        Manifest manifest2 = new Manifest(reactor);
+        Manifest baselineManifest = new Manifest(baseline);
+        Manifest projectManifest = new Manifest(reactor);
 
-        Attributes attributes = manifest.getMainAttributes();
-        Attributes attributes2 = manifest2.getMainAttributes();
+        Attributes baselineAttributes = baselineManifest.getMainAttributes();
+        Attributes projectAttributes = projectManifest.getMainAttributes();
 
         Set<Name> names = new LinkedHashSet<>();
-        names.addAll(getNames(attributes));
-        names.addAll(getNames(attributes2));
+        names.addAll(getNames(baselineAttributes));
+        names.addAll(getNames(projectAttributes));
 
         for (Name key : names) {
-            String value = attributes.getValue(key);
-            if (value == null) {
-                addDelta(result, key, "not present in baseline version");
+            String baselineValue = baselineAttributes.getValue(key);
+            String reactorValue = projectAttributes.getValue(key);
+            if (baselineValue == null) {
+                addDelta(result, key, "not present in baseline version", Change.PROJECT_ONLY, reactorValue,
+                        reactorValue);
                 continue;
             }
 
-            String value2 = attributes2.getValue(key);
-            if (value2 == null) {
-                addDelta(result, key, "present in baseline version only");
+            if (reactorValue == null) {
+                addDelta(result, key, "present in baseline version only", Change.BASELINE_ONLY, baselineValue,
+                        baselineValue);
                 continue;
             }
 
-            if (!value.equals(value2)) {
-                addDelta(result, key, "baseline='" + value + "' != reactor='" + value2 + "'");
+            if (!baselineValue.equals(reactorValue)) {
+                addDelta(result, key, "baseline='" + baselineValue + "' != reactor='" + reactorValue + "'",
+                        Change.DIFFERENT, baselineValue, reactorValue);
             }
         }
-
+        checkForEEChange(result);
         return !result.isEmpty() ? new CompoundArtifactDelta("different", result) : ArtifactDelta.NO_DIFFERENCE;
     }
 
-    private void addDelta(TreeMap<String, ArtifactDelta> result, Name key, String message) {
-        result.put(key.toString(), new SimpleArtifactDelta(message));
+    private void checkForEEChange(Map<String, ManifestDelta> result) {
+        try {
+            if (result.size() > 1) {
+                @SuppressWarnings("deprecation")
+                Entry<String, ManifestDelta> bree = result.entrySet().stream()
+                        .filter(e -> Constants.BUNDLE_REQUIREDEXECUTIONENVIRONMENT.equalsIgnoreCase(e.getKey()))
+                        .findFirst().orElse(null);
+                ManifestDelta breeDelta;
+                if (bree != null && (breeDelta = bree.getValue()).change != Change.DIFFERENT) {
+                    //check if it was migrated to ee.cap...
+                    Entry<String, ManifestDelta> cap = result.entrySet().stream()
+                            .filter(e -> Constants.REQUIRE_CAPABILITY.equalsIgnoreCase(e.getKey())).findFirst()
+                            .orElse(null);
+                    if (cap != null) {
+                        ManifestDelta capDelta = cap.getValue();
+                        if (capDelta.change != Change.DIFFERENT) {
+                            if (isEquivialentBreeCap(breeDelta.value, capDelta.value)) {
+                                result.remove(cap.getKey());
+                                result.remove(bree.getKey());
+                            }
+                            return;
+                        }
+                        if (breeDelta.change == Change.BASELINE_ONLY) {
+                            if (isEquivialentBreeCap(breeDelta.value, capDelta.value, capDelta.changed)) {
+                                result.remove(cap.getKey());
+                                result.remove(bree.getKey());
+                            }
+                        }
+                    }
+                }
+            }
+        } catch (RuntimeException e) {
+            //just in case any invalid/unkownn values we can't process, then we simply  say it is different anyways
+        }
+    }
+
+    private void addDelta(TreeMap<String, ManifestDelta> result, Name key, String message, Change change, String value,
+            String changed) {
+        result.put(key.toString(), new ManifestDelta(message, change, value, changed));
     }
 
     protected Set<Name> getNames(Attributes attributes) {
@@ -97,6 +144,9 @@ public class ManifestComparator implements ContentsComparator {
     }
 
     public static boolean isIgnoredHeaderName(String name) {
+        if (name == null) {
+            return false;
+        }
         return IGNORED_KEYS.contains(new Name(name));
     }
 
@@ -107,5 +157,39 @@ public class ManifestComparator implements ContentsComparator {
     @Override
     public boolean matches(String extension) {
         return TYPE.equalsIgnoreCase(extension);
+    }
+
+    private static final class ManifestDelta extends SimpleArtifactDelta {
+
+        private Change change;
+        private String value;
+        private String changed;
+
+        public ManifestDelta(String message, Change change, String value, String changed) {
+            super(message);
+            this.change = change;
+            this.value = value;
+            this.changed = changed;
+        }
+
+    }
+
+    public static boolean isEquivialentBreeCap(String bree, String cap) {
+        String[] parts = bree.split("-");
+        String osgiee = "osgi.ee;filter:=\"(&(osgi.ee=" + parts[0] + ")(version=" + parts[1] + "))\"";
+        return cap.equalsIgnoreCase(osgiee);
+    }
+
+    public static boolean isEquivialentBreeCap(String bree, String capBase, String capChange) {
+        String[] parts = bree.split("-");
+        String osgiee = "osgi.ee;filter:=\"(&(osgi.ee=" + parts[0] + ")(version=" + parts[1] + "))\"";
+        String withoutEE = capChange.replace(osgiee, "").replace(",,", ",");
+        if (withoutEE.endsWith(",")) {
+            withoutEE = withoutEE.substring(0, withoutEE.length() - 1);
+        }
+        if (withoutEE.startsWith(",")) {
+            withoutEE = withoutEE.substring(1);
+        }
+        return capBase.equalsIgnoreCase(withoutEE);
     }
 }